LayoutTests/fast/frames/seamless/seamless-hyperlink.html

<!DOCTYPE html>
<script src="../../js/resources/js-test-pre.js"></script>
<script>
if (window.testRunner)
    testRunner.waitUntilDone();

window.addEventListener('message', function(evt) {
    shouldBeEqualToString("frames[0].document.body.textContent", "DONE\n");

    if (window.testRunner)
        testRunner.notifyDone();
}, false);
</script>
<iframe srcdoc="
    Waiting...
    <script>
        window.onload = function() {
            frames[0].document.getElementsByTagName('a')[0].click();
        }
    </script>
    <iframe seamless srcdoc='<a href=resources/done.html>Click me</a>'></iframe>
"></iframe>
